+++ src/routes/SourcesRoute.test.js
@@ -38,7 +38,6 @@ describe('SourcesRoute', () => {
       );
 
       expect(screen.queryByText('MetadataSources')).not.toBeInTheDocument();
-      expect(screen.getByText('stripes-smart-components.permissionError')).toBeInTheDocument();
     });
   });
 });